Home work to lesson 6 Decimal and pageble

+++ wnteredshop/api/src/main/java/ru/geekbrains/wnteredshop/api/OrderItemDto.java
@@ -1,6 +1,7 @@
 package ru.geekbrains.wnteredshop.api;
 
 
+import java.math.BigDecimal;
 
 public class OrderItemDto {
 
@@ -13,7 +14,7 @@ public class OrderItemDto {
 
     private int quantity;
 
-    private int pricePerProduct;
+    private BigDecimal pricePerProduct;
 
     public Long getId() {
         return id;
@@ -47,21 +48,21 @@ public class OrderItemDto {
         this.quantity = quantity;
     }
 
-    public int getPricePerProduct() {
+    public BigDecimal getPricePerProduct() {
         return pricePerProduct;
     }
 
-    public void setPricePerProduct(int pricePerProduct) {
+    public void setPricePerProduct(BigDecimal pricePerProduct) {
         this.pricePerProduct = pricePerProduct;
     }
 
-    public int getPrice() {
+    public BigDecimal getPrice() {
         return price;
     }
 
-    public void setPrice(int price) {
+    public void setPrice(BigDecimal price) {
         this.price = price;
     }
 
-    private int price;
+    private BigDecimal price;
 }
